Load javascript modification independently of ACF, closes #1.
This solution overrides acf.relationship_update_results() only,
which previously had to be hacked in the ACF code. This will of course
break if ACF upgrades include significant changes to this method,
but at least it has a chance of surviving multiple ACF releases
and removes the need to modify ACF source directly.